requirements.txt
flask==3.0.3
gunicorn==21.2.0
requests==2.32.3
python-dotenv==1.0.1
groq>=0.9.0
psycopg2-binary==2.9.9